+++ test/index_test.rb
@@ -28,6 +28,11 @@ class IndexTest < Minitest::Test
     assert !old_index.exists?
   end
 
+  def test_total_docs
+    store_names ["Product A"]
+    assert_equal 1, Product.searchkick_index.total_docs
+  end
+
   def test_mapping
     store_names ["Dollar Tree"], Store
     assert_equal [], Store.search(query: {match: {name: "dollar"}}).map(&:name)
@@ -88,18 +93,18 @@ class IndexTest < Minitest::Test
   end
 
   def test_missing_index
-    assert_raises(Searchkick::MissingIndexError) { Product.search "test", index_name: "not_found" }
+    assert_raises(Searchkick::MissingIndexError) { Product.search("test", index_name: "not_found") }
   end
 
   def test_unsupported_version
-    raises_exception = ->(_) { raise Elasticsearch::Transport::Transport::Error.new("[500] No query registered for [multi_match]") }
+    raises_exception = ->(_) { raise Elasticsearch::Transport::Transport::Error, "[500] No query registered for [multi_match]" }
     Searchkick.client.stub :search, raises_exception do
       assert_raises(Searchkick::UnsupportedVersionError) { Product.search("test") }
     end
   end
 
-  def test_invalid_query
-    assert_raises(Searchkick::InvalidQueryError) { Product.search(query: {boom: true}) }
+  def test_invalid_body
+    assert_raises(Searchkick::InvalidQueryError) { Product.search(body: {boom: true}) }
   end
 
   def test_transaction
@@ -112,6 +117,14 @@ class IndexTest < Minitest::Test
   end
 
   def test_analyzed_only
+    # skip for 5.0 since it throws
+    # Cannot search on field [alt_description] since it is not indexed.
+    skip unless elasticsearch_below50?
+    store [{name: "Product A", alt_description: "Hello"}]
+    assert_search "*", [], where: {alt_description: "Hello"}
+  end
+
+  def test_analyzed_only_large_value
     skip if nobrainer?
     large_value = 10000.times.map { "hello" }.join(" ")
     store [{name: "Product A", alt_description: large_value}]