+++ test/test_helper.rb
@@ -25,6 +25,10 @@ def elasticsearch_below50?
   Searchkick.server_below?("5.0.0-alpha1")
 end
 
+def elasticsearch_below22?
+  Searchkick.server_below?("2.2.0")
+end
+
 def elasticsearch_below20?
   Searchkick.server_below?("2.0.0")
 end
@@ -93,6 +97,19 @@ if defined?(Mongoid)
     field :name
   end
 
+  class Region
+    include Mongoid::Document
+
+    field :name
+    field :text
+  end
+
+  class Speaker
+    include Mongoid::Document
+
+    field :name
+  end
+
   class Animal
     include Mongoid::Document
 
@@ -137,6 +154,21 @@ elsif defined?(NoBrainer)
     field :name, type: String
   end
 
+  class Region
+    include NoBrainer::Document
+
+    field :id,   type: Object
+    field :name, type: String
+    field :text, type: Text
+  end
+
+  class Speaker
+    include NoBrainer::Document
+
+    field :id,   type: Object
+    field :name, type: String
+  end
+
   class Animal
     include NoBrainer::Document
 
@@ -162,7 +194,7 @@ else
   # migrations
   ActiveRecord::Base.establish_connection adapter: "sqlite3", database: ":memory:"
 
-  ActiveRecord::Base.raise_in_transactional_callbacks = true if ActiveRecord::Base.respond_to?(:raise_in_transactional_callbacks=)
+  ActiveRecord::Base.raise_in_transactional_callbacks = true if ActiveRecord::VERSION::STRING.start_with?("4.2.")
 
   if defined?(Apartment)
     class Rails
@@ -221,16 +253,32 @@ else
     t.string :name
   end
 
+  ActiveRecord::Migration.create_table :regions do |t|
+    t.string :name
+    t.text :text
+  end
+
+  ActiveRecord::Migration.create_table :speakers do |t|
+    t.string :name
+  end
+
   ActiveRecord::Migration.create_table :animals do |t|
     t.string :name
     t.string :type
   end
 
   class Product < ActiveRecord::Base
+    belongs_to :store
   end
 
   class Store < ActiveRecord::Base
     has_many :products
+  end
+
+  class Region < ActiveRecord::Base
+  end
+
+  class Speaker < ActiveRecord::Base
   end
 
   class Animal < ActiveRecord::Base
@@ -244,8 +292,6 @@ else
 end
 
 class Product
-  belongs_to :store
-
   searchkick \
     synonyms: [
       ["clorox", "bleach"],
@@ -253,13 +299,15 @@ class Product
       ["saranwrap", "plasticwrap"],
       ["qtip", "cottonswab"],
       ["burger", "hamburger"],
-      ["bandaid", "bandag"]
+      ["bandaid", "bandag"],
+      "lightbulb => led,lightbulb",
+      "lightbulb => halogenlamp"
     ],
     autocomplete: [:name],
     suggest: [:name, :color],
-    conversions: "conversions",
-    personalize: "user_ids",
-    locations: ["location", "multiple_locations"],
+    conversions: [:conversions],
+    personalize: :user_ids,
+    locations: [:location, :multiple_locations],
     text_start: [:name],
     text_middle: [:name],
     text_end: [:name],
@@ -267,12 +315,14 @@ class Product
     word_middle: [:name],
     word_end: [:name],
     highlight: [:name],
-    # unsearchable: [:description],
     searchable: [:name, :color],
-    only_analyzed: [:alt_description],
+    default_fields: [:name, :color],
+    filterable: [:name, :color, :description],
+    # unsearchable: [:description],
+    # only_analyzed: [:alt_description],
     match: ENV["MATCH"] ? ENV["MATCH"].to_sym : nil
 
-  attr_accessor :conversions, :user_ids, :aisle
+  attr_accessor :conversions, :user_ids, :aisle, :details
 
   def search_data
     serializable_hash.except("id").merge(
@@ -280,12 +330,19 @@ class Product
       user_ids: user_ids,
       location: {lat: latitude, lon: longitude},
       multiple_locations: [{lat: latitude, lon: longitude}, {lat: 0, lon: 0}],
-      aisle: aisle
+      aisle: aisle,
+      details: details
     )
   end
 
   def should_index?
     name != "DO NOT INDEX"
+  end
+
+  def search_name
+    {
+      name: name
+    }
   end
 end
 
@@ -310,6 +367,38 @@ class Store
   end
 end
 
+class Region
+  searchkick \
+    geo_shape: {
+      territory: {tree: "quadtree", precision: "10km"}
+    }
+
+  attr_accessor :territory
+
+  def search_data
+    {
+      name: name,
+      text: text,
+      territory: territory
+    }
+  end
+end
+
+class Speaker
+  searchkick \
+    conversions: ["conversions_a", "conversions_b"]
+
+  attr_accessor :conversions_a, :conversions_b, :aisle
+
+  def search_data
+    serializable_hash.except("id").merge(
+      conversions_a: conversions_a,
+      conversions_b: conversions_b,
+      aisle: aisle
+    )
+  end
+end
+
 class Animal
   searchkick \
     autocomplete: [:name],
@@ -325,12 +414,15 @@ Product.create!(name: "Set mapping")
 
 Store.reindex
 Animal.reindex
+Speaker.reindex
+Region.reindex
 
 class Minitest::Test
   def setup
     Product.destroy_all
     Store.destroy_all
     Animal.destroy_all
+    Speaker.destroy_all
   end
 
   protected
@@ -353,6 +445,10 @@ class Minitest::Test
 
   def assert_order(term, expected, options = {}, klass = Product)
     assert_equal expected, klass.search(term, options).map(&:name)
+  end
+
+  def assert_equal_scores(term, options = {}, klass = Product)
+    assert_equal 1, klass.search(term, options).hits.map { |a| a["_score"] }.uniq.size
   end
 
   def assert_first(term, expected, options = {}, klass = Product)